# docs/hermes-agent/collection.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hakeem Hermes Agent API",
    "version": "1.0.0",
    "description": "Customer-service API for the Hermes agent. Every request requires X-Hermes-Key. Profile and bookings also need a customer bearer token or phone/email + password. No admin, delete, or clinic-management operations."
  },
  "servers": [
    {
      "url": "{HAKEEM_BASE_URL}/api/agent/v1",
      "description": "Hakeem agent surface",
      "variables": {
        "HAKEEM_BASE_URL": {
          "default": "http://localhost:8000"
        }
      }
    }
  ],
...
  "security": [
    { "hermesKey": [] }
  ],
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "search",
        "tags": ["Catalog"],
        "summary": "Search doctors, clinics, services, offers, labs",
        "parameters": [
          { "$ref": "#/components/parameters/Locale" },
          { "name": "q", "in": "query", "schema": { "type": "string", "maxLength": 120 } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["all", "doctors", "clinics", "services", "offers", "labs"] } },
          { "name": "specialty", "in": "query", "schema": { "type": "string" }, "description": "Specialty slug" },
          { "name": "governorate", "in": "query", "schema": { "type": "string" }, "description": "Governorate slug" },
          { "name": "city", "in": "query", "schema": { "type": "string" }, "description": "City slug" },
          { "name": "gender", "in": "query", "schema": { "type": "string", "enum": ["male", "female"] } },
          { "name": "min_experience", "in": "query", "schema": { "type": "integer", "minimum": 0, "maximum": 70 } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "max_price", "in": "query", "schema": { "type": "integer", "minimum": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 24 } }
        ],
        "responses": { "200": { "description": "Search results" }, "401": { "$ref": "#/components/responses/HermesUnauthorized" } }
      }
    },
...
    "/reference/geography": {
      "get": {
        "operationId": "referenceGeography",
        "tags": ["Catalog"],
        "summary": "Governorates and cities",
        "responses": { "200": { "description": "Geography" }, "401": { "$ref": "#/components/responses/HermesUnauthorized" } }
      }
    },
    "/reference/specialties": {
      "get": {
        "operationId": "referenceSpecialties",
        "tags": ["Catalog"],
        "summary": "Active specialties",
        "responses": { "200": { "description": "Specialties" }, "401": { "$ref": "#/components/responses/HermesUnauthorized" } }
      }
    },
    "/reference/service-types": {
      "get": {
        "operationId": "referenceServiceTypes",
        "tags": ["Catalog"],
        "summary": "Bookable service types",
        "responses": { "200": { "description": "Service types" }, "401": { "$ref": "#/components/responses/HermesUnauthorized" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "hermesKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Hermes-Key",
        "description": "Platform Hermes service key. Required on every request."
      },
      "customerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Patient Sanctum token. Use when this conversation already authenticated the visitor. Do not ask for a password again."
      }
    },
    "parameters": {
      "Locale": {
        "name": "locale",
        "in": "query",
        "schema": { "type": "string", "enum": ["ar", "en"] },
        "description": "Response language. You may send X-Locale instead."
      }
    },
    "responses": {
      "HermesUnauthorized": {
        "description": "Missing or wrong X-Hermes-Key",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "code": { "type": "string", "example": "hermes_key_required" }
              }
            }
          }
        }
      },
...
    },
...
  }
}